+++ R/RcppExports.R
@@ -10,14 +10,16 @@
 #'        expected by PGM/PPM image format) data ordering must be converted. If this argument
 #'        is set to FALSE, then image output will be faster (due to fewer data-ordering operations, and
 #'        better cache coherency) but the image will be transposed. Default: TRUE
+#' @param flipy By default, the position [0, 0] is considered the top-left corner of the output image. Set flipy = TRUE
+#'        for [0, 0] to represent the bottom-left corner. Default: flipy = FALSE
 #' @param intensity_factor Multiplication factor applied to all values in image
 #'        (note: no checking is performed to ensure values remain in range [0, 1]).
 #'        If intensity_factor <= 0, then automatically determine (and apply) a multiplication factor
 #'        to set the maximum value to 1.0. Default: intensity_factor = 1.0
 #'
 #'
-write_pgm <- function(mat, filename, convert_to_row_major = TRUE, intensity_factor = 1) {
-    invisible(.Call(`_foist_write_pgm`, mat, filename, convert_to_row_major, intensity_factor))
+write_pgm <- function(mat, filename, convert_to_row_major = TRUE, flipy = FALSE, intensity_factor = 1) {
+    invisible(.Call(`_foist_write_pgm`, mat, filename, convert_to_row_major, flipy, intensity_factor))
 }
 
 #' Write a numeric matrix to a PPM file using the specified palette
@@ -31,14 +33,16 @@ write_pgm <- function(mat, filename, convert_to_row_major = TRUE, intensity_fact
 #'        expected by PGM/PPM image format) data ordering must be converted. If this argument
 #'        is set to FALSE, then image output will be faster (due to fewer data-ordering operations, and
 #'        better cache coherency) but the image will be transposed. Default: TRUE
+#' @param flipy By default, the position [0, 0] is considered the top-left corner of the output image. Set flipy = TRUE
+#'        for [0, 0] to represent the bottom-left corner. Default: flipy = FALSE
 #' @param intensity_factor Multiplication factor applied to all values in image
 #'        (note: no checking is performed to ensure values remain in range [0, 1]).
 #'        If intensity_factor <= 0, then automatically determine (and apply) a multiplication factor
 #'        to set the maximum value to 1.0. Default: intensity_factor = 1.0
 #'
 #'
-write_pal_ppm <- function(mat, pal, filename, convert_to_row_major = TRUE, intensity_factor = 1) {
-    invisible(.Call(`_foist_write_pal_ppm`, mat, pal, filename, convert_to_row_major, intensity_factor))
+write_pal_ppm <- function(mat, pal, filename, convert_to_row_major = TRUE, flipy = FALSE, intensity_factor = 1) {
+    invisible(.Call(`_foist_write_pal_ppm`, mat, pal, filename, convert_to_row_major, flipy, intensity_factor))
 }
 
 #' Write a numeric array to a PPM file
@@ -53,13 +57,15 @@ write_pal_ppm <- function(mat, pal, filename, convert_to_row_major = TRUE, inten
 #'        expected by PGM/PPM image format) data ordering must be converted. If this argument
 #'        is set to FALSE, then image output will be faster (due to fewer data-ordering operations, and
 #'        better cache coherency) but the image will be transposed. Default: TRUE
+#' @param flipy By default, the position [0, 0] is considered the top-left corner of the output image. Set flipy = TRUE
+#'        for [0, 0] to represent the bottom-left corner. Default: flipy = FALSE
 #' @param intensity_factor Multiplication factor applied to all values in image
 #'        (note: no checking is performed to ensure values remain in range [0, 1]).
 #'        If intensity_factor <= 0, then automatically determine (and apply) a multiplication factor
 #'        to set the maximum value to 1.0. Default: intensity_factor = 1.0
 #'
 #'
-write_ppm_core <- function(arr, dims, filename, convert_to_row_major = TRUE, intensity_factor = 1) {
-    invisible(.Call(`_foist_write_ppm_core`, arr, dims, filename, convert_to_row_major, intensity_factor))
+write_ppm_core <- function(arr, dims, filename, convert_to_row_major = TRUE, flipy = FALSE, intensity_factor = 1) {
+    invisible(.Call(`_foist_write_ppm_core`, arr, dims, filename, convert_to_row_major, flipy, intensity_factor))
 }