+++ NEWS.md
@@ -1,3 +1,11 @@
-# foist 0.1
+
+# foist 0.1.1
+
+* Added `flipy` flag
+* Better handling of column-major/row-major by swapping nrow and ncol 
+* Since image size and buffer size are both known, the process of buffer output
+  to file has been streamlined.
+
+# foist 0.1.0
 
 * Initial Release